drives kinda straight

+++ stm32_f103/Core/Inc/MadgwickAHRS.h
@@ -25,8 +25,8 @@ extern float w_bx, w_by, w_bz;
 //---------------------------------------------------------------------------------------------------
 // Function declarations
 
-void MadgwickAHRSupdate(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz, float dt);
-void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az, float dt);
+void MadgwickAHRSupdate(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz, double dt);
+void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az, double dt);
 void computeAngles();
 float getRoll();
 float getPitch();
